requirements.txt
PyYAML
python-dotenv
discord.py
slack_bolt
types-requests